+++ RELEASES.md
@@ -7,6 +7,7 @@ Notable changes:
  - xrefs.Decorations: only return Reference targets in DecorationsReply.Nodes
  - Use proto3 JSON mapping for web requests: https://developers.google.com/protocol-buffers/docs/proto3#json
  - Java indexer: report error when indexing from compilation's source root
+ - Consistently use corpus root relative paths in filetree API
 
 Notable additions:
  - Java indexer: emit (possibly multi-line) snippets over entire surrounding statement